+++ __integration__/jest-integration/__integration__/steps/test5.steps.ts
@@ -1,4 +1,4 @@
-import { Given, Teardown, When } from "autometa-runner";
+import { Given, Teardown, When } from "@autometa/runner";
 import { expect } from "@jest/globals";
 const things = ["one", "three"];
 const counts = [1, 3];