+++ __integration__/jest-integration/src/apps/default.world.ts
@@ -1,3 +1,3 @@
-import { AutometaWorld } from "autometa-runner";
+import { AutometaWorld } from "@autometa/runner";
 
 export class World extends AutometaWorld {}