+++ packages/jest-transformer/src/index.ts
@@ -28,7 +28,7 @@ export default {
   process: (_src: string, filePath: string, jestConfig: Config.ProjectConfig) => {
     const windowsFix = filePath.replace(/\\/g, "/");
     const testFile = `
-const { Feature } = require('autometa-runner');            
+const { Feature } = require('@autometa/runner');            
 Feature('${windowsFix}')
 `;